deploy resources to TASK_EXEC_RESULT_DATA_BUCKET
