--- csv-to-pdf/src/resources/cars_report.xml
@@ -1,86 +1,0 @@
-<?xml version = "1.0" encoding = "UTF-8"?>
-<!DOCTYPE jasperReport PUBLIC "//JasperReports//DTD Report Design//EN"
-        "http://jasperreports.sourceforge.net/dtds/jasperreport.dtd">
-
-<jasperReport xmlns="http://jasperreports.sourceforge.net/jasperreports"
-              xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-              xsi:schemaLocation="http://jasperreports.sourceforge.net/jasperreports
-   http://jasperreports.sourceforge.net/xsd/jasperreport.xsd"
-              name="report2" topMargin="20" bottomMargin="20">
-
-    <field name="Id" class="java.lang.Long">
-        <fieldDescription><![CDATA[id]]></fieldDescription>
-    </field>
-
-    <field name="Name">
-        <fieldDescription><![CDATA[name]]></fieldDescription>
-    </field>
-
-    <field name="Price" class="java.lang.Integer">
-        <fieldDescription><![CDATA[price]]></fieldDescription>
-    </field>
-
-    <sortField name="Id"/>
-
-    <columnHeader>
-        <band height="15">
-
-            <staticText>
-                <reportElement x="0" y="0" width="300" height="15"/>
-                <textElement />
-                <text><![CDATA[]]> </text>
-            </staticText>
-
-            <staticText>
-                <reportElement x="0" y="0" width="50" height="15"/>
-                <textElement />
-                <text><![CDATA[Id]]></text>
-            </staticText>
-
-            <staticText>
-                <reportElement x="80" y="0" width="100" height="15"/>
-                <textElement />
-                <text><![CDATA[Name]]></text>
-            </staticText>
-
-            <staticText>
-                <reportElement x="150" y="0" width="100" height="15"/>
-                <textElement textAlignment="Center"/>
-                <text><![CDATA[Price]]></text>
-            </staticText>
-
-        </band>
-    </columnHeader>
-
-    <detail>
-        <band height="15">
-
-            <textField>
-                <reportElement x="0" y="0" width="50" height="15"/>
-                <textElement/>
-                <textFieldExpression class="java.lang.Long">
-                    <![CDATA[$F{Id}]]>
-                </textFieldExpression>
-            </textField>
-
-            <textField>
-                <reportElement x="80" y="0" width="100" height="15"/>
-                <textElement textAlignment="Left"/>
-                <textFieldExpression>
-                    <![CDATA[$F{Name}]]>
-                </textFieldExpression>
-            </textField>
-
-            <textField>
-                <reportElement x="150" y="0" width="100" height="15"/>
-                <textElement textAlignment="Right"/>
-                <textFieldExpression class="java.lang.Integer">
-                    <![CDATA[$F{Price}]]>
-                </textFieldExpression>
-            </textField>
-
-        </band>
-    </detail>
-
-</jasperReport>
-
